Automatic_Baby_Vein_Detection
Automatic_Baby_Vein_Detection:

In any surgery, the first important phase is an insertion of an intravenous (IV) catheter. The major problem face by the physicians is difficulty in accessing vein for IV drug delivery or taking blood samples for test. In case of children, adults, critical care patients etc. It is very difficult task. Unnecessary puncturing of veins occurs due to poor visibility. Wrong puncturing may leads many problems such as bruises or permanently damage the vein. To overcome from these problems there is a need of vein detection system. The paper presents an approach for live baby Vein detection System. In the proposed work live camera is used to identify the veins. The preprocessing is done to enhance the vein image using CLAHE.